+++ descope/authmethod/_sso_base.py
@@ -22,6 +22,11 @@ class SSOBase:
             raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Tenant cannot be empty")
 
     @staticmethod
+    def _validate_exchange_code(code: str) -> None:
+        if not code:
+            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "exchange code is empty")
+
+    @staticmethod
     def _compose_start_params(
         tenant: str,
         return_url: str,
