+++ descope/authmethod/enchantedlink_async.py
@@ -24,8 +24,7 @@ class EnchantedLinkAsync(EnchantedLinkBase, AsyncAuthBase):
         login_options: LoginOptions | None = None,
         refresh_token: str | None = None,
     ) -> dict:
-        if not login_id:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "login_id is empty")
+        self._validate_sign_in_login_id(login_id)
 
         validate_refresh_token_provided(login_options, refresh_token)
 
@@ -93,8 +92,7 @@ class EnchantedLinkAsync(EnchantedLinkBase, AsyncAuthBase):
         template_id: str | None = None,
         provider_id: str | None = None,
     ) -> dict:
-        if not login_id:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Identifier cannot be empty")
+        self._validate_login_id(login_id)
 
         Auth.validate_email(email)
