+++ descope/authmethod/magiclink.py
@@ -25,12 +25,7 @@ class MagicLink(MagicLinkBase, AuthBase):
         login_options: LoginOptions | None = None,
         refresh_token: str | None = None,
     ) -> str:
-        if not login_id:
-            raise AuthException(
-                400,
-                ERROR_TYPE_INVALID_ARGUMENT,
-                "Identifier is empty",
-            )
+        self._validate_sign_in_login_id(login_id)
 
         validate_refresh_token_provided(login_options, refresh_token)
 
@@ -106,8 +101,7 @@ class MagicLink(MagicLinkBase, AuthBase):
         template_id: str | None = None,
         provider_id: str | None = None,
     ) -> str:
-        if not login_id:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Identifier cannot be empty")
+        self._validate_login_id(login_id)
 
         Auth.validate_email(email)
 
@@ -136,8 +130,7 @@ class MagicLink(MagicLinkBase, AuthBase):
         template_id: str | None = None,
         provider_id: str | None = None,
     ) -> str:
-        if not login_id:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Identifier cannot be empty")
+        self._validate_login_id(login_id)
 
         Auth.validate_phone(method, phone)
