+++ descope/authmethod/password_async.py
@@ -5,7 +5,6 @@ from typing import Iterable
 from descope._auth_base import AsyncAuthBase
 from descope.authmethod._password_base import PasswordBase
 from descope.common import REFRESH_SESSION_COOKIE_NAME, EndpointsV1
-from descope.exceptions import ERROR_TYPE_INVALID_ARGUMENT, AuthException
 
 
 class PasswordAsync(PasswordBase, AsyncAuthBase):
@@ -18,11 +17,8 @@ class PasswordAsync(PasswordBase, AsyncAuthBase):
         user: dict | None = None,
         audience: str | None | Iterable[str] = None,
     ) -> dict:
-        if not login_id:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "login_id cannot be empty")
-
-        if not password:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "password cannot be empty")
+        self._validate_login_id(login_id)
+        self._validate_password(password)
 
         uri = EndpointsV1.sign_up_password_path
         body = self._compose_signup_body(login_id, password, user)
@@ -37,11 +33,8 @@ class PasswordAsync(PasswordBase, AsyncAuthBase):
         password: str,
         audience: str | None | Iterable[str] = None,
     ) -> dict:
-        if not login_id:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "login_id cannot be empty")
-
-        if not password:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Password cannot be empty")
+        self._validate_login_id(login_id)
+        self._validate_sign_in_password(password)
 
         uri = EndpointsV1.sign_in_password_path
         response = await self._http.post(uri, body={"loginId": login_id, "password": password})
@@ -55,8 +48,7 @@ class PasswordAsync(PasswordBase, AsyncAuthBase):
         redirect_url: str | None = None,
         template_options: dict | None = None,
     ) -> dict:
-        if not login_id:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "login_id cannot be empty")
+        self._validate_login_id(login_id)
 
         uri = EndpointsV1.send_reset_password_path
         body: dict[str, str | bool | dict | None] = {
@@ -70,14 +62,9 @@ class PasswordAsync(PasswordBase, AsyncAuthBase):
         return response.json()
 
     async def update(self, login_id: str, new_password: str, refresh_token: str) -> None:
-        if not login_id:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "login_id cannot be empty")
-
-        if not new_password:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "new_password cannot be empty")
-
-        if not refresh_token:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Refresh token cannot be empty")
+        self._validate_login_id(login_id)
+        self._validate_new_password(new_password)
+        self._validate_refresh_token(refresh_token)
 
         uri = EndpointsV1.update_password_path
         await self._http.post(
@@ -93,14 +80,9 @@ class PasswordAsync(PasswordBase, AsyncAuthBase):
         new_password: str,
         audience: str | None | Iterable[str] = None,
     ) -> dict:
-        if not login_id:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "login_id cannot be empty")
-
-        if not old_password:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "old_password cannot be empty")
-
-        if not new_password:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "new_password cannot be empty")
+        self._validate_login_id(login_id)
+        self._validate_old_password(old_password)
+        self._validate_new_password(new_password)
 
         uri = EndpointsV1.replace_password_path
         response = await self._http.post(
