+++ descope/authmethod/webauthn.py
@@ -10,7 +10,6 @@ from descope.common import (
     LoginOptions,
     validate_refresh_token_provided,
 )
-from descope.exceptions import ERROR_TYPE_INVALID_ARGUMENT, AuthException
 
 
 class WebAuthn(WebAuthnBase, AuthBase):
@@ -23,11 +22,8 @@ class WebAuthn(WebAuthnBase, AuthBase):
         """
         Docs
         """
-        if not login_id:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Identifier cannot be empty")
-
-        if not origin:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Origin cannot be empty")
+        self._validate_login_id(login_id)
+        self._validate_origin(origin)
 
         if not user:
             user = {}
@@ -46,11 +42,8 @@ class WebAuthn(WebAuthnBase, AuthBase):
         """
         Docs
         """
-        if not transaction_id:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Transaction id cannot be empty")
-
-        if not response:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Response cannot be empty")
+        self._validate_transaction_id(transaction_id)
+        self._validate_response(response)
         uri = EndpointsV1.sign_up_auth_webauthn_finish_path
         body = self._compose_sign_up_in_finish_body(transaction_id, response)
         response = self._http.post(uri, body=body)
@@ -67,11 +60,8 @@ class WebAuthn(WebAuthnBase, AuthBase):
         """
         Docs
         """
-        if not login_id:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Identifier cannot be empty")
-
-        if not origin:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Origin cannot be empty")
+        self._validate_login_id(login_id)
+        self._validate_origin(origin)
 
         validate_refresh_token_provided(login_options, refresh_token)
 
@@ -89,11 +79,8 @@ class WebAuthn(WebAuthnBase, AuthBase):
         """
         Docs
         """
-        if not transaction_id:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Transaction id cannot be empty")
-
-        if not response:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Response cannot be empty")
+        self._validate_transaction_id(transaction_id)
+        self._validate_response(response)
 
         uri = EndpointsV1.sign_in_auth_webauthn_finish_path
         body = self._compose_sign_up_in_finish_body(transaction_id, response)
@@ -109,11 +96,8 @@ class WebAuthn(WebAuthnBase, AuthBase):
         """
         Docs
         """
-        if not login_id:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Identifier cannot be empty")
-
-        if not origin:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Origin cannot be empty")
+        self._validate_login_id(login_id)
+        self._validate_origin(origin)
 
         uri = EndpointsV1.sign_up_or_in_auth_webauthn_start_path
         body = self._compose_sign_up_or_in_start_body(login_id, origin)
@@ -124,11 +108,8 @@ class WebAuthn(WebAuthnBase, AuthBase):
         """
         Docs
         """
-        if not login_id:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Identifier cannot be empty")
-
-        if not refresh_token:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Refresh token cannot be empty")
+        self._validate_login_id(login_id)
+        self._validate_refresh_token(refresh_token)
 
         uri = EndpointsV1.update_auth_webauthn_start_path
         body = self._compose_update_start_body(login_id, origin)
@@ -139,11 +120,8 @@ class WebAuthn(WebAuthnBase, AuthBase):
         """
         Docs
         """
-        if not transaction_id:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Transaction id cannot be empty")
-
-        if not response:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Response cannot be empty")
+        self._validate_transaction_id(transaction_id)
+        self._validate_response(response)
 
         uri = EndpointsV1.update_auth_webauthn_finish_path
         body = self._compose_update_finish_body(transaction_id, response)
