+++ descope/authmethod/webauthn_async.py
@@ -10,7 +10,6 @@ from descope.common import (
     LoginOptions,
     validate_refresh_token_provided,
 )
-from descope.exceptions import ERROR_TYPE_INVALID_ARGUMENT, AuthException
 
 
 class WebAuthnAsync(WebAuthnBase, AsyncAuthBase):
@@ -22,11 +21,8 @@ class WebAuthnAsync(WebAuthnBase, AsyncAuthBase):
         origin: Optional[str],
         user: Optional[dict] = None,
     ) -> dict:
-        if not login_id:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Identifier cannot be empty")
-
-        if not origin:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Origin cannot be empty")
+        self._validate_login_id(login_id)
+        self._validate_origin(origin)
 
         if not user:
             user = {}
@@ -42,11 +38,8 @@ class WebAuthnAsync(WebAuthnBase, AsyncAuthBase):
         response,
         audience: Union[str, None, Iterable[str]] = None,
     ) -> dict:
-        if not transaction_id:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Transaction id cannot be empty")
-
-        if not response:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Response cannot be empty")
+        self._validate_transaction_id(transaction_id)
+        self._validate_response(response)
 
         uri = EndpointsV1.sign_up_auth_webauthn_finish_path
         body = self._compose_sign_up_in_finish_body(transaction_id, response)
@@ -61,11 +54,8 @@ class WebAuthnAsync(WebAuthnBase, AsyncAuthBase):
         login_options: Optional[LoginOptions] = None,
         refresh_token: Optional[str] = None,
     ) -> dict:
-        if not login_id:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Identifier cannot be empty")
-
-        if not origin:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Origin cannot be empty")
+        self._validate_login_id(login_id)
+        self._validate_origin(origin)
 
         validate_refresh_token_provided(login_options, refresh_token)
 
@@ -80,11 +70,8 @@ class WebAuthnAsync(WebAuthnBase, AsyncAuthBase):
         response,
         audience: Union[str, None, Iterable[str]] = None,
     ) -> dict:
-        if not transaction_id:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Transaction id cannot be empty")
-
-        if not response:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Response cannot be empty")
+        self._validate_transaction_id(transaction_id)
+        self._validate_response(response)
 
         uri = EndpointsV1.sign_in_auth_webauthn_finish_path
         body = self._compose_sign_up_in_finish_body(transaction_id, response)
@@ -97,11 +84,8 @@ class WebAuthnAsync(WebAuthnBase, AsyncAuthBase):
         login_id: str,
         origin: str,
     ) -> dict:
-        if not login_id:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Identifier cannot be empty")
-
-        if not origin:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Origin cannot be empty")
+        self._validate_login_id(login_id)
+        self._validate_origin(origin)
 
         uri = EndpointsV1.sign_up_or_in_auth_webauthn_start_path
         body = self._compose_sign_up_or_in_start_body(login_id, origin)
@@ -109,11 +93,8 @@ class WebAuthnAsync(WebAuthnBase, AsyncAuthBase):
         return response.json()
 
     async def update_start(self, login_id: str, refresh_token: str, origin: str) -> dict:
-        if not login_id:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Identifier cannot be empty")
-
-        if not refresh_token:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Refresh token cannot be empty")
+        self._validate_login_id(login_id)
+        self._validate_refresh_token(refresh_token)
 
         uri = EndpointsV1.update_auth_webauthn_start_path
         body = self._compose_update_start_body(login_id, origin)
@@ -121,11 +102,8 @@ class WebAuthnAsync(WebAuthnBase, AsyncAuthBase):
         return response.json()
 
     async def update_finish(self, transaction_id: str, response: str) -> None:
-        if not transaction_id:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Transaction id cannot be empty")
-
-        if not response:
-            raise AuthException(400, ERROR_TYPE_INVALID_ARGUMENT, "Response cannot be empty")
+        self._validate_transaction_id(transaction_id)
+        self._validate_response(response)
 
         uri = EndpointsV1.update_auth_webauthn_finish_path
         body = self._compose_update_finish_body(transaction_id, response)
